requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
cffi==1.15.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
h11==0.14.0
idna==3.4
iso8601==1.1.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pypika-tortoise==0.1.6
pytz==2022.7
sniffio==1.3.0
starlette==0.22.0
tortoise-orm==0.19.2
typing_extensions==4.4.0
uvicorn==0.20.0
